refactor(core/messages): move extract_last_assistant_text from diagnosis
Transcript parsing belongs with message helpers; repoint diagnose node and
tests to core.messages.transcript.
