+++ core/domain/diagnosis/__init__.py
@@ -9,7 +9,6 @@ from core.domain.diagnosis.result import (
     build_diagnosis_schema,
     build_investigation_result,
     claims_to_dicts,
-    extract_last_assistant_text,
     normalize_root_cause_category,
     result_to_state,
     root_cause_category_instruction_for_source,
@@ -23,7 +22,6 @@ __all__ = [
     "build_investigation_result",
     "claims_to_dicts",
     "detect_category_text_mismatch",
-    "extract_last_assistant_text",
     "normalize_root_cause_category",
     "result_to_state",
     "root_cause_category_instruction_for_source",
