+++ core/domain/diagnosis/result.py
@@ -97,33 +97,6 @@ def result_to_state(result: InvestigationResult) -> dict[str, Any]:
     }
 
 
-def extract_last_assistant_text(messages: list[dict[str, Any]]) -> str:
-    for msg in reversed(messages):
-        if msg.get("role") != "assistant":
-            continue
-        content = msg.get("content", "")
-        if isinstance(content, str) and content.strip():
-            return content.strip()
-        if isinstance(content, list):
-            parts: list[str] = []
-            for block in content:
-                if isinstance(block, str):
-                    parts.append(block)
-                    continue
-                if isinstance(block, dict):
-                    if block.get("type") == "text" and isinstance(block.get("text"), str):
-                        parts.append(block["text"])
-                    continue
-                block_type = getattr(block, "type", None)
-                block_text = getattr(block, "text", None)
-                if block_type == "text" and isinstance(block_text, str):
-                    parts.append(block_text)
-            text = " ".join(p for p in parts if p).strip()
-            if text:
-                return text
-    return ""
-
-
 def taxonomy_categories_for_alert_source(alert_source: str) -> set[str]:
     source = alert_source.strip().lower()
     if source == "hermes":
